requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
alabaster==1.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
apeye==1.4.1
apeye-core==1.1.5
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
ARLBench @ git+https://github.com/stamchry/arlbench.git@main#egg=ARLBench
arrow==1.4.0
astroid==4.0.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
autodocsumm==0.2.14
automl_sphinx_theme==0.2.0
babel==2.17.0
backports.tarfile==1.2.0
backrefs==5.9
beautifulsoup4==4.14.2
black==25.9.0
bleach==6.3.0
blinker==1.9.0
bracex==2.6
brax==0.10.3
bump-my-version==1.2.4
CacheControl==0.14.3
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chex==0.1.86
click==8.2.1
cloudpickle==3.1.1
codecarbon==2.3.5
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
ConfigSpace==1.2.1
contourpy==1.3.2
coverage==7.4.4
cryptography==46.0.3
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
dask==2025.10.0
dask-jobqueue==0.9.0
debugpy==1.8.17
decopatch==1.4.10
decorator==5.2.1
defusedxml==0.7.1
dict2css==0.3.0.post1
dill==0.4.0
distlib==0.4.0
distrax==0.1.5
distributed==2025.10.0
dm-env==1.6
dm-tree==0.1.9
docker-pycreds==0.4.0
docutils==0.21.2
domdf_python_tools==3.10.0
emcee==3.1.6
etils==1.13.0
exceptiongroup==1.3.0
execnet==2.1.1
executing==2.2.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.20.0
flake8==7.3.0
flashbax==0.1.2
Flask==3.1.2
flask-cors==6.0.1
flax==0.8.2
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
griffe==1.14.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.0
gymnax==0.0.6
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
HyperSweeper @ git+https://github.com/stamchry/hypersweeper.git@main#egg=HyperSweeper
identify==2.6.15
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==7.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jax==0.4.26
jaxlib==0.4.26
jaxopt==0.8.5
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.9
lark==1.3.1
linkify-it-py==2.0.3
locket==1.0.0
makefun==1.16.0
Markdown==3.9
markdown-exec==1.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
memray==1.12.0
mergedeep==1.3.4
mike==2.1.3
mistune==3.1.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-gen-files==0.5.0
mkdocs-get-deps==0.2.0
mkdocs-glightbox==0.5.2
mkdocs-glossary-plugin==0.1.1
mkdocs-literate-nav==0.6.2
mkdocs-material==9.6.22
mkdocs-material-extensions==1.3.1
mkdocstrings==0.30.1
mkdocstrings-python==1.18.2
ml_collections==1.1.0
ml_dtypes==0.5.3
more-itertools==10.8.0
msgpack==1.1.2
mujoco==3.1.3
mujoco-mjx==3.1.3
multiprocess==0.70.18
mypy==1.18.2
mypy_extensions==1.1.0
myst-parser==4.0.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.3.1
nodeenv==1.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
numpydoc==1.9.0
nvidia-ml-py==13.580.82
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.2
orbax-checkpoint==0.4.4
overrides==7.7.0
packaging==25.0
paginate==0.5.7
pandas==2.1.4
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.5
partd==1.4.2
pathos==0.3.2
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
pkginfo==1.10.0
platformdirs==4.5.0
pluggy==1.6.0
plumbum==1.9.0
ply==3.11
pox==0.3.6
ppft==1.7.7
pre-commit==3.7.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
pygments-ansi-color==0.3.0
pylint==4.0.2
pymdown-extensions==10.16.1
pynisher==1.0.10
pynvml==13.0.1
PyOpenGL==3.1.10
pyparsing==3.2.5
pyrfr==0.9.0
pytest==7.4.4
pytest-cases==3.9.1
pytest-cov==4.1.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytinyrenderer==0.0.14
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
pyzmq==27.1.0
qtconsole==5.7.0
QtPy==2.4.3
questionary==2.1.1
RapidFuzz==3.14.1
readme_renderer==44.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-click==1.9.4
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
ruff==0.14.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.4.0
selectolax==0.4.0
Send2Trash==1.8.3
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smac @ git+https://github.com/stamchry/SMAC3.git@main#egg=smac
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
Sphinx==8.1.3
sphinx-autodoc-typehints==3.0.1
sphinx-gallery==0.19.0
sphinx-jinja2-compat==0.4.1
sphinx-prompt==1.9.0
sphinx-tabs==3.4.5
sphinx-toolbox==4.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
submitit==1.5.3
tabulate==0.9.0
tblib==3.2.0
tensorboardX==2.6.4
tensorflow-probability==0.25.0
tensorstore==0.1.78
terminado==0.18.1
textual==6.4.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.13.3
toolz==1.1.0
tornado==6.5.2
traitlets==5.14.3
trimesh==4.9.0
twine==5.1.1
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.5.0
uv==0.9.6
verspec==0.1.0
virtualenv==20.35.4
wandb==0.16.6
watchdog==6.0.0
wcmatch==10.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==2.0.0
xminigrid==0.8.0
zict==3.0.0
zipp==3.23.0
